:recycle: issue #2 UIカード以外DDDアーキテクチャに基づき、リファクタリング完了
